refactor: halo counter resource name
